+++ repomanager/app.py
@@ -2,7 +2,6 @@
 # Entry point for the application.
 from . import create_app
 
-app = create_app()
-
 if __name__ == "__main__":
+    app = create_app()
     app.run()
