+++ repomanager/templates/user.html
@@ -69,7 +69,7 @@
   <h3>Repositories</h3>
   <!-- repositories -->
   <p>
-    There are {{user.repos|length}} <a href="{{url_for('views.repos',username=user.login)}}">repositories</a> in total.
+    There are {{user.repos|length}} <a href="{{url_for('user.repos',username=user.login)}}">repositories</a> in total.
   </p>
   <ul>
     {% for repo in user.repos %}
