+++ repomanager/views/__init__.py
@@ -1,0 +1,5 @@
+from .user import user_blueprint
+from .public import public_blueprint
+from .auth import auth_blueprint
+
+__all__ = ["user_blueprint", "public_blueprint", "auth_blueprint"]
